Split internal/operator/operator.go into multiple files
